+++ packages/odata-v4/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-cloud-sdk/odata-v4
 
+## 3.15.0
+
+### Patch Changes
+
+- @sap-cloud-sdk/connectivity@3.15.0
+- @sap-cloud-sdk/http-client@3.15.0
+- @sap-cloud-sdk/odata-common@3.15.0
+- @sap-cloud-sdk/util@3.15.0
+
 ## 3.14.0
 
 ### Patch Changes