+++ test-packages/self-tests/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @sap-cloud-sdk/self-tests
 
+## 3.15.0
+
 ## 3.14.0
 
 ## 3.13.0